backend/requirements.txt
pillow>=10.0
imageio>=2.34
imageio-ffmpeg>=0.4
numpy>=1.26
